+++ spark-cypher/src/main/scala/org/opencypher/spark/impl/convert/SparkConversions.scala
@@ -35,7 +35,7 @@ import org.opencypher.okapi.api.value.CypherValue.{CypherMap, CypherValue, Cyphe
 import org.opencypher.okapi.impl.exception.{IllegalArgumentException, NotImplementedException}
 import org.opencypher.okapi.ir.api.expr.Var
 import org.opencypher.okapi.relational.impl.table.RecordHeader
-import org.opencypher.spark.impl.temporal.SparkTemporalHelpers._
+import org.opencypher.spark.impl.temporal.TemporalConversions._
 import org.opencypher.spark.impl.SparkSQLMappingException
 
 object SparkConversions {
